requirements.txt
Flask==3.1.1
transformers==4.53.0
torch==2.7.1+cpu
pillow==11.2.1
numpy==2.2.6
werkzeug==3.1.3
huggingface-hub==0.33.1
safetensors==0.5.3
tqdm==4.67.1
regex==2024.11.6
requests==2.32.3
python-dotenv==1.1.0
